web/server/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
Pillow
aiofiles
